+++ index.html
@@ -6,7 +6,7 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script defer src="/js/main.js"></script>
+  <script defer src="/js/main.js" type="module"></script>
   <title>Кекстаграм</title>
 </head>
