+++ js/const.js
@@ -1,0 +1,33 @@
+// Массив описаний фото
+const DESCRIPTION = [
+  'Дисплей',
+  'Окружающая среда',
+  'Приоритет',
+  'Гордость образования',
+  'Романтика',
+  'Смелость',
+  'Поражение штормов',
+  'Источник тайны',
+  'Осень',
+  'Неудачное падение',
+  'Уверенность в культуре',
+  'Дым блеска',
+  'Влияние',
+  'Последствия',
+  'Древние трофеи',
+  'Бизнес',
+  'Забавный контракт',
+  'Внешность',
+  'Приоритет жизни',
+  'Шторм',
+  'Город эмоций',
+  'Проблемы реальности',
+  'Сюрприз образования',
+  'Благотворительность',
+  'Покинутое здравомыслие'
+];
+export {DESCRIPTION};
+
+//Количество создаваемых объектов
+const COUNT_CARDS_PHOTO = 25;
+export {COUNT_CARDS_PHOTO};
